CI: Minor refactoring in ci_utils
