+++ tests/ci/ci_config.py
@@ -3,7 +3,7 @@ import re
 from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 from typing import Dict, Optional, List
 
-from ci_utils import normalize_string
+from ci_utils import Utils
 from ci_definitions import *
 
 
@@ -13,7 +13,6 @@ class CI:
     each config item in the below dicts should be an instance of JobConfig class or inherited from it
     """
 
-    MAX_TOTAL_FAILURES_BEFORE_BLOCKING_CI = 5
     MAX_TOTAL_FAILURES_PER_JOB_BEFORE_BLOCKING_CI = 2
 
     # reimport types to CI class so that they visible as CI.* and mypy is happy
@@ -37,9 +36,7 @@ class CI:
     from ci_utils import GH as GH
     from ci_utils import Shell as Shell
     from ci_definitions import Labels as Labels
-    from ci_definitions import TRUSTED_CONTRIBUTORS as TRUSTED_CONTRIBUTORS
     from ci_definitions import WorkFlowNames as WorkFlowNames
-    from ci_utils import CATEGORY_TO_LABEL as CATEGORY_TO_LABEL
 
     # Jobs that run for doc related updates
     _DOCS_CHECK_JOBS = [JobNames.DOCS_CHECK, JobNames.STYLE_CHECK]
@@ -558,7 +555,7 @@ class CI:
     @classmethod
     def get_tag_config(cls, label_name: str) -> Optional[LabelConfig]:
         for label, config in cls.TAG_CONFIGS.items():
-            if normalize_string(label_name) == normalize_string(label):
+            if Utils.normalize_string(label_name) == Utils.normalize_string(label):
                 return config
         return None
